REGISTRATION_CP/requirements.txt
# ---- Core Django ----
Django
djangorestframework
django-cors-headers
django-environ
psycopg2-binary 
django-countries


# ---- Auth & Accounts ----
dj-rest-auth
django-allauth
djangorestframework-simplejwt
social-auth-app-django

#---- Google Signup/Login ----

google-auth
google-auth-oauthlib 
google-auth-httplib2

# ---- Django Admin Enhancements ----
django-jazzmin
django-import-export
django-simple-history
whitenoise

# ---- Payments ----
stripe

# ---- LiveKit Integration ----
livekit
livekit-api
livekit-protocol

# ---- Celery & Redis ----
celery
redis
django-celery-results
flower

# ---- Security & Validation ----
cryptography
django-recaptcha
python-dotenv
bcrypt
python-decouple

# ---- Utility ----
requests
aiohttp
aiofiles
Pillow
humanize

# ---- Optional (Admin / Parsing Helpers) ----
beautifulsoup4
lxml
python-magic
django-ratelimit

# ---- Required for typing / data models ----
pydantic
pydantic-settings
annotated-types

# ---- Task retry helpers / scheduling ----
backoff
schedule

# ---- File type identification ----
filetype
moviepy
ffmpeg
